refactor(tests): update import paths for platform services and entities
